+++ changelogs/mediation/startapp/CHANGELOG.md
@@ -1,6 +1,12 @@
 # Change Log
 All notable changes to StartApp Adapter for Yandex Mobile Ads Mediation will be documented in this file.
 
+## Version 5.0.2.0
+
+#### Updated
+* Yandex Mobile Ads SDK version 7.3.0
+* Startapp SDK version 5.0.2
+
 ## Version 5.0.1.0
 
 #### Updated